backend/notesapp/requirements.txt
Django==1.11
djangorestframework==3.6.2
django-cors-headers==2.0.2
